pyproject.toml
[project]
name = "test-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "google-cloud-pubsub>=2.29.0",
    "uv>=0.7.8",
]
